First committ
very and i mean VERY basic wa-tor envirnoment. working fish (labeled as F) and sharks (S), in an otherwise empty space ( . ). The Fish and Sharks move randomly in steps sperated by a unit of time. Behaviour of sharks yet to be implemented, and the GUI, and the Concurrency part also yet to be implemented. i.e a lot yet to be implement